Need another :Delview that is more extreme. Hence :DelviewHard
Delview works except on BufWinLeave, mkview is called again.
The specific problem was when I

>$ vimdiff ~/.vimrc vimrc

cursorbind gets set and saved in the view

then even if I

>$ vim ~/.vimrc

and :Delview, cursorbind is still set in the buffer
so when I :q, on BufWinLeave, mkview is called, and cursorbind is
preserved.

So :DelviewHard is :Delview | noautocmd :q